got this to actually work. how nice
