+++ backend/src/lib/selection-committee.ts
@@ -3,7 +3,8 @@ import { parse } from 'csv-parse/sync';
 import fs from 'fs';
 import path from 'path';
 import { CollegeNameFormatter } from './college-name-formatter';
-import { SelectionCommitteeType, MasseyEndpointsType, MasseyEndpointsByYearType, TeamDataType, GameTeamDataType, GameDataType, TeamSummaryDataType, YearRankingsType, TeamDataMap, TeamSummaryMap } from './types';
+import { SelectionCommitteeType, MasseyEndpointsType, MasseyEndpointsByYearType, TeamDataType, GameTeamDataType, GameDataType, TeamSummaryDataType, YearRankingsType, TeamDataMap, TeamSummaryMap, TeamSummaryWithProbabilityType } from './types';
+import { MBBLogisticalRegression } from './logistical-regression';
 
 const mbbEndpointsByYear = _mbbEndpointsByYear as MasseyEndpointsByYearType;
 const nameTable:{ [key: string]: string } = require(path.join(__dirname, '../../data/name-table.json'));
@@ -91,7 +92,7 @@ export class SelectionCommittee implements SelectionCommitteeType {
 		return tournamentStartDate;
 	}
 
-	async getTeamNames(year: number) {
+	async getTeamNames(year: number):Promise<Map<string,string>> {
 		const endpoints = this.getMasseyEndpoints(year);
 		const response = await fetch(endpoints.teams);
 		const data = await response.text();
@@ -313,8 +314,10 @@ export class SelectionCommittee implements SelectionCommitteeType {
 		
 	}
 
-	cleanTeamNames(teams: any) {
-		const cleanedTeams = teams.map((team: any) => {
+	cleanTeamNames(teams: any):Map<string,string> {
+		
+
+		const cleanedTeams: {id: string, name: string}[] = teams.map((team: any) => {
 			return {
 				'id' : team[0].trim(),
 				'name' : SelectionCommittee.getPrettyTeamName(team[1].trim())
@@ -399,5 +402,40 @@ export class SelectionCommittee implements SelectionCommitteeType {
 			return 'road';
 		}
 		
+	}
+
+
+
+	async getTeamSummariesWithProbabilities():Promise<TeamSummaryWithProbabilityType[]> {
+		const dataDir = path.join(__dirname, '../../data/archive');
+		
+		const teamSummaries = this.getTeamSummariesFromFile();
+		const finalSummaries:TeamSummaryWithProbabilityType[] = [];
+
+		const logisticalRegression = new MBBLogisticalRegression();
+		logisticalRegression.loadModel();
+
+		const teamNames = await this.getTeamNames(this.year);
+
+		teamSummaries.forEach((teamSummary:TeamSummaryDataType) => {
+			const result = logisticalRegression.predictProbability(teamSummary);
+			const teamName:string = teamNames.get(teamSummary.id.toString()) || '[unknown]';
+
+			const finalSummary:TeamSummaryWithProbabilityType = {
+				...teamSummary,
+				teamName: teamName,
+				probability: result.probability,
+				prediction: result.prediction,
+				rawDecisionValue: result.rawDecisionValue
+			}
+
+			finalSummaries.push(finalSummary);
+			
+		});
+
+		fs.writeFileSync(path.join(dataDir, `mensbb-team-summaries-${this.year}-with-probabilities.json`), JSON.stringify(finalSummaries, null, 2));
+
+
+		return finalSummaries;
 	}
 }
